pyproject.toml
[tool.poetry]
name = "repl_python3_Minecraftserverapi"
version = "0.1.0"
description = ""
authors = ["OpGamer3000Pro <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
replit = "^3.2.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
